+++ webknossos/webknossos/cli/convert_zarr.py
@@ -13,6 +13,9 @@ import typer
 import zarr
 from typing_extensions import Annotated
 
+from webknossos.dataset.length_unit import LengthUnit
+from webknossos.dataset.properties import DEFAULT_LENGTH_UNIT_STR, VoxelSize
+
 from ..dataset import DataFormat, Dataset, MagView, SegmentationLayer
 from ..dataset._array import _fsstore_from_path
 from ..dataset.defaults import DEFAULT_CHUNK_SHAPE, DEFAULT_CHUNKS_PER_SHARD
@@ -21,7 +24,7 @@ from ..utils import get_executor_for_args, wait_and_ensure_success
 from ._utils import (
     DistributionStrategy,
     SamplingMode,
-    VoxelSize,
+    VoxelSizeTuple,
     parse_mag,
     parse_path,
     parse_vec3int,
@@ -60,7 +63,7 @@ def convert_zarr(
     chunk_shape: Vec3Int,
     chunks_per_shard: Vec3Int,
     is_segmentation_layer: bool = False,
-    voxel_size: Optional[Tuple[float, float, float]] = (1.0, 1.0, 1.0),
+    voxel_size_with_unit: VoxelSize = VoxelSize((1.0, 1.0, 1.0)),
     flip_axes: Optional[Union[int, Tuple[int, ...]]] = None,
     compress: bool = True,
     executor_args: Optional[Namespace] = None,
@@ -72,9 +75,9 @@ def convert_zarr(
     input_dtype: Any = file.dtype
     shape: Any = file.shape
 
-    if voxel_size is None:
-        voxel_size = 1.0, 1.0, 1.0
-    wk_ds = Dataset(target_path, voxel_size=voxel_size, exist_ok=True)
+    wk_ds = Dataset(
+        target_path, voxel_size_with_unit=voxel_size_with_unit, exist_ok=True
+    )
     wk_layer = wk_ds.get_or_add_layer(
         layer_name,
         "segmentation" if is_segmentation_layer else "color",
@@ -139,14 +142,21 @@ def main(
         typer.Option(help="Name of the cubed layer (color or segmentation)"),
     ] = "color",
     voxel_size: Annotated[
-        Optional[VoxelSize],
+        VoxelSizeTuple,
         typer.Option(
             help="The size of one voxel in source data in nanometers. "
             "Should be a comma separated string (e.g. 11.0,11.0,20.0).",
             parser=parse_voxel_size,
-            metavar="VOXEL_SIZE",
+            metavar="VoxelSize",
+            show_default=False,
         ),
-    ] = None,
+    ],
+    unit: Annotated[
+        LengthUnit,
+        typer.Option(
+            help="The unit of the voxel size.",
+        ),
+    ] = DEFAULT_LENGTH_UNIT_STR,  # type:ignore
     data_format: Annotated[
         DataFormat,
         typer.Option(
@@ -243,6 +253,7 @@ When converting a folder, this option is ignored."
         distribution_strategy=distribution_strategy.value,
         job_resources=job_resources,
     )
+    voxel_size_with_unit = VoxelSize(voxel_size, unit)
 
     mag_view = convert_zarr(
         source,
@@ -252,7 +263,7 @@ When converting a folder, this option is ignored."
         chunk_shape=chunk_shape,
         chunks_per_shard=chunks_per_shard,
         is_segmentation_layer=is_segmentation_layer,
-        voxel_size=voxel_size,
+        voxel_size_with_unit=voxel_size_with_unit,
         flip_axes=flip_axes,
         compress=compress,
         executor_args=executor_args,